Fix macOS Clang OpenMP support: Use -Xpreprocessor -fopenmp in Makefile. Cast mpz_sizeinbase to int in epsilon_tuning.c to resolve sign warning. Verified build and run: outputs ε=0.2500.

+++ src/c/grok-terminal/python/src/c/rsa/epsilon_tuning.c
@@ -16,7 +16,7 @@ void generate_random_prime(mpz_t prime, gmp_randstate_t state, int bits) {
         mpz_urandomb(rand_num, state, bits - 1);
         mpz_setbit(rand_num, bits - 1);  // Ensure it's at least 2^(bits-1)
         mpz_nextprime(prime, rand_num);
-    } while (mpz_sizeinbase(prime, 2) > bits);  // Cap at bits
+    } while ((int)mpz_sizeinbase(prime, 2) > bits);  // Cap at bits (cast to avoid warning)
     mpz_clear(rand_num);
 }
